# src/main/java/br/edu/fiec/MapeamentoDeSaude/features/auth/controllers/AuthController.java
package br.edu.fiec.MapeamentoDeSaude.features.auth.controllers;


import br.edu.fiec.MapeamentoDeSaude.features.auth.dto.LoginRequest;
import br.edu.fiec.MapeamentoDeSaude.features.auth.dto.RegisterRequest;
import br.edu.fiec.MapeamentoDeSaude.features.auth.services.AuthService;
import br.edu.fiec.MapeamentoDeSaude.features.user.models.User;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/v1/api/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/register")
    public ResponseEntity<User> register(@RequestBody RegisterRequest request) {
        User newUser = authService.register(request);
        return new ResponseEntity<>(newUser, HttpStatus.CREATED);
    }

    @PostMapping("/login")
    public ResponseEntity<User> login(@RequestBody LoginRequest request) {
        User loggedInUser = authService.login(request);
        return new ResponseEntity<>(loggedInUser, HttpStatus.OK);
    }
}